DHFPROD-4857: hubApply fails immediately if file is not found

+++ marklogic-data-hub/src/main/java/com/marklogic/hub/hubcentral/HubCentralManager.java
@@ -66,6 +66,10 @@ public class HubCentralManager extends LoggingObject {
      * @param zipFile
      */
     public void applyHubCentralZipToProject(HubProject hubProject, File zipFile) {
+        if (zipFile == null || !zipFile.exists()) {
+            throw new RuntimeException("Unable to apply zip file to project, file does not exist: " + zipFile);
+        }
+
         deleteUserArtifacts(hubProject);
         deleteEntityBasedArtifacts(hubProject);
